requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
requests>=2.26.0
python-multipart>=0.0.5
psycopg2-binary>=2.9.9
pydantic>=2.6.4
